+++ aasrp/__init__.py
@@ -5,5 +5,5 @@ App init
 # Django
 from django.utils.translation import gettext_lazy as _
 
-__version__ = "2.6.1"
+__version__ = "2.7.0"
 __title__ = _("Ship Replacement")