+++ CHANGELOG.md
@@ -41,6 +41,10 @@ Section Order:
 ### Security
 -->
 
+### Changed
+
+- Use framework CSS classes (Bootstrap and AA Framework) instead of custom CSS
+
 ## [2.6.1] - 2025-03-08
 
 ### Changed