+++ CHANGELOG.md
@@ -41,6 +41,14 @@ Section Order:
 ### Security
 -->
 
+### Changed
+
+- Use framework CSS classes (Bootstrap and AA Framework) instead of custom CSS
+
+### Removed
+
+- Unused code
+
 ## [2.6.1] - 2025-03-08
 
 ### Changed